moved dialog components to dialog folder
