--- client/src/components/Buttons/AddCardButton/useStyles.ts
@@ -1,14 +1,0 @@
-import { makeStyles } from '@material-ui/core/styles';
-
-const useStyles = makeStyles(() => ({
-  addCardButton: {
-    color: 'white',
-    paddingTop: 5,
-    paddingBottom: 5,
-    paddingLeft: 20,
-    paddingRight: 20,
-    backgroundColor: '#759CFC',
-  },
-}));
-
-export default useStyles;
